+++ test/serialize_with_alternate_handlers_test.dart
@@ -12,8 +12,8 @@ void main() {
                 Parameter1: 42,
                 otherParameter: 'Value',
                 innerClass: DummyClass()),
-            [],
-            UsePropertiesOnMissingSerializationInformation()),
+            fallbackObjectSerialization:
+                UsePropertiesOnMissingSerializationInformation()),
         'O:19:"ClassWithParameters":3:{s:10:"Parameter1";i:42;s:14:"otherParameter";s:5:"Value";s:10:"innerClass";O:10:"DummyClass":0:{}}');
   });
   test('Serialize class with automatic property mapping - including getters',
@@ -24,9 +24,9 @@ void main() {
                 Parameter1: 42,
                 otherParameter: 'Value',
                 innerClass: DummyClass()),
-            [],
-            UsePropertiesOnMissingSerializationInformation(
-                inspectGetters: true)),
+            fallbackObjectSerialization:
+                UsePropertiesOnMissingSerializationInformation(
+                    inspectGetters: true)),
         'O:19:"ClassWithParameters":4:{s:10:"Parameter1";i:42;s:14:"otherParameter";s:5:"Value";s:10:"innerClass";O:10:"DummyClass":0:{}s:18:"parameterViaGetter";i:42;}');
   });
   test(
@@ -38,9 +38,9 @@ void main() {
                 Parameter1: 42,
                 otherParameter: 'Value',
                 innerClass: DummyClass()),
-            [],
-            UsePropertiesOnMissingSerializationInformation(
-                inspectPrivate: true)),
+            fallbackObjectSerialization:
+                UsePropertiesOnMissingSerializationInformation(
+                    inspectPrivate: true)),
         'O:19:"ClassWithParameters":4:{s:10:"Parameter1";i:42;s:14:"otherParameter";s:5:"Value";s:10:"innerClass";O:10:"DummyClass":0:{}s:16:"_hiddenDuplicate";s:5:"Value";}');
   });
 }